# Conflicts:
#	src/main/java/com/example/end/controller/ProcedureController.java
#	src/main/java/com/example/end/controller/UserController.java
#	src/main/java/com/example/end/dto/BookingDto.java
#	src/main/java/com/example/end/dto/CartDto.java
#	src/main/java/com/example/end/dto/ProcedureDto.java
#	src/main/java/com/example/end/models/Booking.java
#	src/main/java/com/example/end/models/Cart.java
#	src/main/java/com/example/end/models/Procedure.java
#	src/main/java/com/example/end/models/Review.java
#	src/main/java/com/example/end/models/Role.java
#	src/main/java/com/example/end/models/User.java
#	src/main/java/com/example/end/models/dto/UserDto.java
#	src/main/java/com/example/end/repository/CartRepository.java
#	src/main/java/com/example/end/repository/ProcedureRepository.java
#	src/main/java/com/example/end/repository/UserRepository.java
#	src/main/java/com/example/end/service/CartServiceImpl.java
#	src/main/java/com/example/end/service/ProcedureServiceImpl.java
#	src/main/java/com/example/end/service/UserServiceImpl.java
#	src/main/java/com/example/end/service/interfaces/ProcedureService.java
#	src/main/java/com/example/end/service/interfaces/UserService.java
#	src/main/resources/application.yml

+++ src/main/java/com/example/end/repository/RoleRepository.java
@@ -4,7 +4,6 @@ package com.example.end.repository;
 import com.example.end.models.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
   Role findByName(String name);
 }
-
